README: complete rewrite to make the argument more solid
The current text of the website is very poorly argued, likely because it
was written as a spur-of-the-moment rant. This is understandable, but if
we want to be taken seriously we need much clearer arguments. This is an
attempt to do exactly that -- by rigorously explaining each of the
problems with this legislation.

Due to the massive change to the text, I have moved all of the
signatories under a section stating that they may not be in favour of
the new text. But if they wish, they can move themselves to the main
signatories section.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>